// frontend/src/shared/lib/api/model/kioskScanResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Cellar
 * OpenAPI spec version: 0.1.0
 */
import type { KioskScanResponseBorrowerOrgName } from './kioskScanResponseBorrowerOrgName';
import type { KioskScanResponseDueDate } from './kioskScanResponseDueDate';

export interface KioskScanResponse {
  plate_id: string;
  barcode: string;
  plate_label: string;
  loan_id: string;
  item_id: string;
  item_status: string;
  action: string;
  borrower_org_id: string;
  borrower_org_name: KioskScanResponseBorrowerOrgName;
  due_date: KioskScanResponseDueDate;
}